Added _utils module in spikesort.
